started restructuring code. moved power sequence code to its own file and created headers to better organize variables
